requirements.txt
django
django-allauth
gunicorn
django-on-heroku
requests
djangorestframework